+++ api/controllers/orders/index.js
@@ -1,0 +1,12 @@
+const { addOrder } = require('./addOrder');
+const { updateOrder } = require('./updateOrder');
+const { getAllOrders } = require('./getAllOrders');
+const { getOrder } = require('./getOrder');
+
+
+module.exports = {
+  addOrder,
+  updateOrder,
+  getAllOrders,
+  getOrder
+};
